test: consolidated all tests into integration folder
