--- gravitee-apim-cypress/cypress/bulk/cypress-config.json
@@ -1,6 +1,0 @@
-{
-  "baseUrl": "http://localhost:8083",
-  "integrationFolder": "cypress/bulk",
-  "testFiles": "**/*.ts",
-  "video": false
-}
