+++ gravitee-apim-cypress/cypress/integration/bulk/cypress-config.json
@@ -1,0 +1,19 @@
+{
+  "baseUrl": "http://localhost:8083",
+  "integrationFolder": "cypress/integration/bulk",
+  "testFiles": "**/*.ts",
+  "video": false,
+  "env": {
+    "failOnStatusCode": false,
+    "api_publisher_user_login": "api1",
+    "api_publisher_user_password": "api1",
+    "application_user_login": "application1",
+    "application_user_password": "application1",
+    "low_permission_user_login": "user",
+    "low_permission_user_password": "password",
+    "admin_user_login": "admin",
+    "admin_user_password": "admin",
+    "managementApi": "/management/organizations/DEFAULT/environments/DEFAULT",
+    "portalApi": "/portal/environments/DEFAULT"
+  }
+}
