+++ gravitee-apim-cypress/cypress/integration/e2e/ui-api-list.spec.ts
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ADMIN_USER, API_PUBLISHER_USER } from '../fakers/users/users';
-import { ApiFakers } from '../fakers/apis';
-import { Api } from '../model/apis';
-import { ApiAssertions } from '../assertions/api.assertion';
-import { gio } from '../commands/gravitee.commands';
+import { ADMIN_USER, API_PUBLISHER_USER } from '../../fakers/users/users';
+import { ApiFakers } from '../../fakers/apis';
+import { Api } from '../../model/apis';
+import { ApiAssertions } from '../../assertions/api.assertion';
+import { gio } from '../../commands/gravitee.commands';
 
 describe('API List feature', () => {
   let createdApi: Api;
