+++ gravitee-apim-cypress/cypress/integration/nrt/apis/publishing.spec.ts
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ADMIN_USER, API_PUBLISHER_USER, APPLICATION_USER, LOW_PERMISSION_USER } from '../../fakers/users/users';
-import { ApiFakers } from '../../fakers/apis';
-import { Api, ApiErrorCodes, ApiLifecycleState, PortalApi } from '../../model/apis';
-import { CollectionResponse, PortalError } from '../../model/technical';
-import { PortalApiAssertions, ApiAssertions } from '../../assertions/api.assertion';
-import { ErrorAssertions } from '../../assertions/error.assertion';
-import { gio } from '../../commands/gravitee.commands';
+import { ADMIN_USER, API_PUBLISHER_USER, APPLICATION_USER, LOW_PERMISSION_USER } from '../../../fakers/users/users';
+import { ApiFakers } from '../../../fakers/apis';
+import { Api, ApiErrorCodes, ApiLifecycleState, PortalApi } from '../../../model/apis';
+import { CollectionResponse, PortalError } from '../../../model/technical';
+import { PortalApiAssertions, ApiAssertions } from '../../../assertions/api.assertion';
+import { ErrorAssertions } from '../../../assertions/error.assertion';
+import { gio } from '../../../commands/gravitee.commands';
 
 context('API - Publishing', () => {
   describe('Not published', () => {
