+++ gravitee-apim-cypress/package.json
@@ -1,36 +1,35 @@
 {
-  "name": "gravitee-apim-cypress",
-  "version": "3.13.0-SNAPSHOT",
-  "private": true,
-  "description": "Gravitee.io APIM - Cypress",
-  "scripts": {
-    "test": "cypress open",
-    "test:bulk": "cypress run --config-file cypress/bulk/cypress-config.json",
-    "test:ci": "cypress run",
-    "test:e2e": "cypress open --config-file cypress/e2e/cypress-config.json",
-    "test:dev": "cypress run --env printLogsToConsole=always",
-    "test:parallel": "cypress run --record --group apis --spec 'cypress/nrt/apis/**/*'",
-    "lint": "echo 'Lint to be configured'",
-    "lint:fix": "npm run prettier:fix && npm run lint:license:fix",
-    "lint:license": "license-check-and-add check -f license-check-config.json",
-    "lint:license:fix": "license-check-and-add add -f license-check-config.json -r",
-    "prettier": "prettier --check \"**/*.{js,ts,html,css,scss,json}\"",
-    "prettier:fix": "prettier --write \"**/*.{js,ts,html,css,scss,json}\"",
-    "prettier:list": "prettier --list-different \"**/*.{js,ts,html,css,scss,json}\""
-  },
-  "license": "MIT License",
-  "devDependencies": {
-    "@types/faker": "5.5.9",
-    "@types/node": "14.17.34",
-    "ansi-regex": "6.0.1",
-    "cypress": "8.7.0",
-    "cypress-terminal-report": "3.4.0",
-    "har-validator": "5.1.5",
-    "license-check-and-add": "4.0.3",
-    "prettier": "2.4.1",
-    "typescript": "4.5.2"
-  },
-  "dependencies": {
-    "faker": "5.5.3"
-  }
+    "name": "gravitee-apim-cypress",
+    "version": "3.13.0-SNAPSHOT",
+    "private": true,
+    "description": "Gravitee.io APIM - Cypress",
+    "scripts": {
+        "test:bulk": "cypress run --config-file cypress/integration/bulk/cypress-config.json",
+        "test:ci": "cypress run --config-file cypress/integration/nrt/cypress-config.json",
+        "test:e2e": "cypress open --config-file cypress/integration/e2e/cypress-config.json",
+        "test:dev": "cypress run --env printLogsToConsole=always --config-file cypress/integration/nrt/cypress-config.json",
+        "test:parallel": "cypress run --record --group apis --spec 'cypress/integration/nrt/apis/**/*' --config-file cypress/integration/nrt/cypress-config.json",
+        "lint": "echo 'Lint to be configured'",
+        "lint:fix": "npm run prettier:fix && npm run lint:license:fix",
+        "lint:license": "license-check-and-add check -f license-check-config.json",
+        "lint:license:fix": "license-check-and-add add -f license-check-config.json -r",
+        "prettier": "prettier --check \"**/*.{js,ts,html,css,scss,json}\"",
+        "prettier:fix": "prettier --write \"**/*.{js,ts,html,css,scss,json}\"",
+        "prettier:list": "prettier --list-different \"**/*.{js,ts,html,css,scss,json}\""
+    },
+    "license": "MIT License",
+    "devDependencies": {
+        "@types/faker": "5.5.9",
+        "@types/node": "14.17.34",
+        "ansi-regex": "6.0.1",
+        "cypress": "8.7.0",
+        "cypress-terminal-report": "3.4.0",
+        "har-validator": "5.1.5",
+        "license-check-and-add": "4.0.3",
+        "prettier": "2.4.1",
+        "typescript": "4.5.2"
+    },
+    "dependencies": {
+        "faker": "5.5.3"
+    }
 }
